# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FácilID — demonstração local",
    "version": "2.1.0",
    "description": "Rotas administrativas exigem X-Admin-Token. Foto capturada e assinatura desenhada ficam cifradas no servidor; o cartão mantém apenas hashes, sem reconhecimento facial ou verificação de autoria. Biometria do aparelho protege uma credencial local; o servidor não recebe digital nem aceita um booleano como prova. CPF + cartão gera desafio; PIN ou credencial de aparelho previamente registrada confirma o acesso. Cartões anteriores sem coleta/PIN exigem reemissão. Use dados fictícios ou voluntários com consentimento."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token"
      }
    },
    "schemas": {
...
      "Emissao": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "cpf",
          "nome",
          "idade",
          "modo",
          "assinatura",
          "pin"
        ],
        "properties": {
          "cpf": {
            "type": "string",
            "pattern": "^[0-9]{11}$",
            "example": "12345678900"
          },
          "nome": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "example": "Maria Silva"
          },
          "idade": {
            "type": "integer",
            "minimum": 0,
            "maximum": 130,
            "example": 72
          },
          "modo": {
            "type": "string",
            "enum": [
              "real",
              "demonstracao"
            ]
          },
          "fotoId": {
            "type": "string",
            "format": "uuid"
          },
          "assinatura": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "largura",
              "altura",
              "tracos"
            ],
            "properties": {
              "largura": {
                "type": "integer",
                "enum": [
                  320
                ]
              },
              "altura": {
                "type": "integer",
                "enum": [
                  180
                ]
              },
              "tracos": {
                "type": "array",
                "minItems": 1,
                "maxItems": 32,
                "description": "Até 1500 pontos no total; pelo menos um segmento com deslocamento após arredondar para 2 casas.",
                "items": {
                  "type": "array",
                  "minItems": 2,
                  "maxItems": 512,
                  "items": {
                    "type": "object",
                    "additionalProperties": false,
                    "required": [
                      "x",
                      "y"
                    ],
                    "properties": {
                      "x": {
                        "type": "number",
                        "minimum": 0,
                        "maximum": 320
                      },
                      "y": {
                        "type": "number",
                        "minimum": 0,
                        "maximum": 180
                      }
                    }
                  }
                }
              }
            }
          },
          "pin": {
            "type": "string",
            "pattern": "^[0-9]{6}$",
            "writeOnly": true
          },
          "consentimento": {
            "type": "boolean"
          }
        },
        "description": "No modo real, fotoId pendente e consentimento:true são obrigatórios. Demonstração usa imagem artificial e não aceita fotoId. PIN não integra o cartão e só é persistido com scrypt/salt dentro do índice cifrado."
      },
...
      "Erro": {
        "type": "object",
        "properties": {
          "mensagem": {
            "type": "string"
          },
          "sucesso": {
            "type": "boolean",
            "enum": [
              false
            ],
            "description": "Presente nas recusas da primeira etapa de autenticação."
          },
          "tentarEm": {
            "type": "integer",
            "format": "int64",
            "description": "Retomada após limite de fatores, em Unix epoch milissegundos."
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "description": "Correlação opcional; também disponível no cabeçalho X-Request-Id"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/emissao": {
      "post": {
        "summary": "Emitir cartão com foto, assinatura desenhada e PIN",
        "description": "Toda emissão recebe UUID novo. Cartão anterior, suas sessões e credenciais de aparelho ficam inválidos. Corpo limitado a 64 KB; RSA/canonicalizador preservados.",
        "security": [
          {
            "adminToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Emissao"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Novo cartão assinado",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chip"
                }
              }
            }
          },
          "400": {
            "description": "Cadastro inválido",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Chave administrativa ausente ou incorreta",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "413": {
            "description": "Corpo maior que o limite do endpoint",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Falha interna ou de persistência; não indica credencial inválida",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/emissao/foto": {
      "post": {
        "summary": "Enviar foto privada de uma emissão",
        "description": "Autorização administrativa é conferida antes do parser de 3 MB. Imagem decodificada limitada a 2 MB, formato/dimensões verificados. Até 50 pendências; limpeza de expiradas ao iniciar ou operar o repositório.",
        "security": [
          {
            "adminToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Foto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Referência de uso único",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenciaFoto"
                }
              }
            }
          },
          "400": {
            "description": "Formato inválido",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Chave administrativa ausente ou incorreta",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "413": {
            "description": "Foto ou corpo grande demais",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "429": {
            "description": "Muitas fotos pendentes",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Falha interna ou de persistência; não indica credencial inválida",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/eventos": {
      "get": {
        "summary": "Consultar histórico administrativo de cartões",
        "security": [
          {
            "adminToken": []
          }
        ],
        "description": "Eventos de emissão, bloqueio e substituição gravados atomicamente com o estado do cartão. Mais recentes primeiro, por ocorridoEm e, em empate, pela ordem inversa de gravação. Não aceita parâmetros de consulta. Repetir bloqueio ou substituir cartão já substituído não duplica eventos. Dados anteriores à implantação ficam sem eventos retroativos. Motivos são definidos pelo servidor; não inclui operador individual, PIN, credenciais ou mídias.",
        "responses": {
          "200": {
            "description": "Eventos administrativos",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventoAdministrativo"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Consulta inválida",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Chave administrativa ausente ou incorreta",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Falha interna ou de persistência; não indica credencial inválida",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/usuarios": {
      "get": {
        "summary": "Listar histórico de emissões sem credenciais assinadas",
        "security": [
          {
            "adminToken": []
          }
        ],
        "description": "Sem parâmetros, preserva o retorno ResumoCartao[]. Para paginar, envie pagina e limite juntos, uma única vez cada, como inteiros positivos seguros escritos somente com dígitos, sem zeros iniciais. Limite máximo 100. Parâmetros desconhecidos, repetidos, parciais ou malformados retornam 400; a autorização administrativa precede a validação da consulta. A ordem original é preservada e páginas além do total retornam itens vazios. Atualmente a listagem completa é fatiada em memória; para volumes maiores, mover contagem e paginação para SQL.",
        "parameters": [
          {
            "in": "query",
            "name": "pagina",
            "required": false,
            "description": "Obrigatório junto com limite; primeira página é 1.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "example": 1
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "description": "Obrigatório junto com pagina; no máximo 100 cartões por página.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "example": 20
          }
        ],
        "responses": {
          "200": {
            "description": "Array sem consulta; objeto paginado quando pagina e limite são enviados",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ResumoCartao"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/PaginaCartoes"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Consulta inválida",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Chave administrativa ausente ou incorreta",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Falha interna ou de persistência; não indica credencial inválida",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/cartoes/{emissaoId}": {
      "get": {
        "summary": "Responsável recupera cartão ativo para a demonstração",
        "security": [
          {
            "adminToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "emissaoId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cartão ativo",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chip"
                }
              }
            }
          },
          "400": {
            "description": "Identificador inválido",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Chave administrativa ausente ou incorreta",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Cartão não encontrado",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Cartão bloqueado ou substituído",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Falha interna ou de persistência; não indica credencial inválida",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/cartoes/{emissaoId}/bloquear": {
      "post": {
        "summary": "Bloquear cartão e revogar suas sessões",
        "description": "Operação idempotente; cartão substituído permanece substituído. Sessões são recusadas na próxima chamada protegida.",
        "security": [
          {
            "adminToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "emissaoId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estado atual",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "emissaoId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "estado": {
                      "type": "string",
                      "enum": [
                        "ativo",
                        "bloqueado",
                        "substituido"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Identificador inválido",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Chave administrativa ausente ou incorreta",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Cartão não encontrado",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "413": {
            "description": "Corpo maior que o limite do endpoint",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Falha interna ou de persistência; não indica credencial inválida",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/cartoes/{emissaoId}/coleta": {
      "get": {
        "summary": "Consultar somente metadados da coleta",
        "security": [
          {
            "adminToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "emissaoId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Metadados privados",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coleta"
                }
              }
            }
          },
          "400": {
            "description": "Identificador inválido",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Chave administrativa ausente ou incorreta",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Coleta não encontrada",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "500": {
            "description": "Falha interna ou de persistência; não indica credencial inválida",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação, sem dados pessoais.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}